# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 拼单平台接口
  version: 0.2.0
  description: |
    T001：框架元数据与健康检查。
    T002：后台身份权限、商品图片、商品库存管理、小程序商品展示。
    金额一律整数分；数量为十进制字符串（≤3 位小数）；时间 ISO 8601 UTC；
    分页 page 从 1 起，pageSize 默认 10、上限 50；响应包络 {data, requestId}，错误 {code, message, requestId, details?}。
servers:
  - url: http://127.0.0.1:3000
    description: 本地联调
paths:
...
  /api/admin/v1/cs/accounts:
    get:
      operationId: listCsAccounts
      summary: "超级管理员读取客服账号（不返回密码）"
      security: [{ bearerAuth: [] }]
      responses:
        '200': {"description":"超级管理员读取客服账号（不返回密码）","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/CsAccount"}}}},"requestId":{"type":"string"}}}}}}
        '400': { $ref: '#/components/responses/Error400' }
        '401': { $ref: '#/components/responses/Error401' }
        '404': { $ref: '#/components/responses/Error404' }
    post:
      operationId: createCsAccount
      summary: "超级管理员创建客服或主管账号"
      security: [{ bearerAuth: [] }]
      requestBody: {"required":true,"content":{"application/json":{"schema":{"type":"object","required":["username","displayName","password","role"],"properties":{"username":{"type":"string"},"displayName":{"type":"string"},"password":{"type":"string","minLength":10,"maxLength":128,"writeOnly":true},"role":{"type":"string","enum":["cs_agent","cs_supervisor"]}}}}}}
      responses:
        '201': {"description":"超级管理员创建客服或主管账号","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/CsAccount"},"requestId":{"type":"string"}}}}}}
        '400': { $ref: '#/components/responses/Error400' }
        '401': { $ref: '#/components/responses/Error401' }
        '404': { $ref: '#/components/responses/Error404' }
  /api/admin/v1/cs/accounts/{id}/status:
    post:
      operationId: updateCsAccountStatus
      summary: "超级管理员启停客服账号，停用撤销全部旧会话"
      security: [{ bearerAuth: [] }]
      parameters: [{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}]
      requestBody: {"required":true,"content":{"application/json":{"schema":{"type":"object","required":["status"],"properties":{"status":{"type":"string","enum":["active","disabled"]}}}}}}
      responses:
        '201': {"description":"超级管理员启停客服账号，停用撤销全部旧会话","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/CsAccount"},"requestId":{"type":"string"}}}}}}
        '400': { $ref: '#/components/responses/Error400' }
        '401': { $ref: '#/components/responses/Error401' }
        '404': { $ref: '#/components/responses/Error404' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 管理员会话 token（admin 域）
    miniUserAuth:
      type: http
      scheme: bearer
      description: 小程序用户会话 token（user 域，与 admin 不可互用）
  responses:
...
    Error400:
      description: 参数校验失败
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Error401:
      description: 未认证
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Error404:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string }
        details:
          type: array
          items: { type: string }
...
    CsAccount: {"type":"object","properties":{"id":{"type":"string","format":"uuid"},"username":{"type":"string"},"displayName":{"type":"string"},"role":{"type":"string","enum":["cs_agent","cs_supervisor"]},"status":{"type":"string","enum":["active","disabled"]}}}